tectonic/xetex-ext.[ch]: rename from XeTeX_ext.[ch]
