+++ build.rs
@@ -150,11 +150,11 @@ fn main() {
         .file("tectonic/dpx-type1c.c")
         .file("tectonic/dpx-unicode.c")
         .file("tectonic/dpx-vf.c")
-        .file("tectonic/XeTeX_ext.c")
         .file("tectonic/xetexini.c")
         .file("tectonic/XeTeX_pic.c")
         .file("tectonic/xetex-engine-interface.c")
         .file("tectonic/xetex-errors.c")
+        .file("tectonic/xetex-ext.c")
         .file("tectonic/xetex-io.c")
         .file("tectonic/xetex-linebreak.c")
         .file("tectonic/xetex-math.c")
